# perf(widgets): avoid unnecessary work when rendering paragraphs