--- a/ratatui/benches/main/paragraph.rs
+++ b/ratatui/benches/main/paragraph.rs
@@ -8,7 +8,6 @@ use ratatui::{
 /// because the scroll offset is a u16, the maximum number of lines that can be scrolled is 65535.
 /// This is a limitation of the current implementation and may be fixed by changing the type of the
 /// scroll offset to a u32.
-const MAX_SCROLL_OFFSET: u16 = u16::MAX;
 const NO_WRAP_WIDTH: u16 = 200;
 const WRAP_WIDTH: u16 = 100;
 
@@ -17,9 +16,10 @@ const WRAP_WIDTH: u16 = 100;
 /// as well as comparing with the various settings on the scroll and wrap features.
 fn paragraph(c: &mut Criterion) {
     let mut group = c.benchmark_group("paragraph");
-    for line_count in [64, 2048, MAX_SCROLL_OFFSET] {
+    for line_count in [64, 2048, u16::MAX] {
         let lines = random_lines(line_count);
         let lines = lines.as_str();
+        let y_scroll = line_count - PARAGRAPH_DEFAULT_HEIGHT;
 
         // benchmark that measures the overhead of creating a paragraph separately from rendering
         group.bench_with_input(BenchmarkId::new("new", line_count), lines, |b, lines| {
@@ -36,14 +36,14 @@ fn paragraph(c: &mut Criterion) {
         // scroll the paragraph by half the number of lines and render
         group.bench_with_input(
             BenchmarkId::new("render_scroll_half", line_count),
-            &Paragraph::new(lines).scroll((0, line_count / 2)),
+            &Paragraph::new(lines).scroll((y_scroll / 2, 0)),
             |bencher, paragraph| render(bencher, paragraph, NO_WRAP_WIDTH),
         );
 
         // scroll the paragraph by the full number of lines and render
         group.bench_with_input(
             BenchmarkId::new("render_scroll_full", line_count),
-            &Paragraph::new(lines).scroll((0, line_count)),
+            &Paragraph::new(lines).scroll((y_scroll, 0)),
             |bencher, paragraph| render(bencher, paragraph, NO_WRAP_WIDTH),
         );
 
@@ -59,16 +59,18 @@ fn paragraph(c: &mut Criterion) {
             BenchmarkId::new("render_wrap_scroll_full", line_count),
             &Paragraph::new(lines)
                 .wrap(Wrap { trim: false })
-                .scroll((0, line_count)),
+                .scroll((y_scroll, 0)),
             |bencher, paragraph| render(bencher, paragraph, WRAP_WIDTH),
         );
     }
     group.finish();
 }
 
+const PARAGRAPH_DEFAULT_HEIGHT: u16 = 50;
+
 /// render the paragraph into a buffer with the given width
 fn render(bencher: &mut Bencher, paragraph: &Paragraph, width: u16) {
-    let mut buffer = Buffer::empty(Rect::new(0, 0, width, 50));
+    let mut buffer = Buffer::empty(Rect::new(0, 0, width, PARAGRAPH_DEFAULT_HEIGHT));
     // We use `iter_batched` to clone the value in the setup function.
     // See https://github.com/ratatui/ratatui/pull/377.
     bencher.iter_batched(