Hi,

Thanks for the great work on ratatui 🙏🏻 

### Context
_This PR introduces a slight change that I believe could end up significantly improving rendering times for paragraphs that do not depend on certain kinds of `LineComposer`s._

Currently all `LineComposer`s are considered to be state machines which means rendering a paragraph with a given Y offset requires computing the entire state up to Y before being able to render from Y onwards.

While this makes sense for Composers such as the `WordWrapper` (where one needs to consider all previous lines to determine where a given line will end up), it means it also penalizes Composers which can render a given line "statelessely" (such as the `LineTruncator`) which actually end up doing a lot of unnecessary work (and on the critical rendering path) when the offset gets high.

I came across this while experimenting with previews in [television](https://github.com/alexpasmantier/television) and noticed increased rendering lag as the preview offset got higher for large files.

#### Example
Below is an example of sampling the application and scrolling down a pretty large file in the preview pane while everything else remains static. Note how rendering the paragraph takes longer and longer (in proportion to the rest which I'm assuming shouldn't change much).


https://github.com/user-attachments/assets/1c6dc428-5351-4659-97c1-45fe7fb2ffb4



### Description of the PR

I tried to keep the PR as minimal as possible. The `LineComposer` trait was split into:
- `StatefulLineComposer`: e.g. `WordWrapper`
- `LineComposer`: e.g. `LineTruncator`

When rendering a paragraph, the behavior is now as follows:
- if the line composer is stateful, use the existing code to compute lines up to offset `y` and render from there on up to the area's height
- if the line composer is stateless, just skip ahead until `y` and render from there on up to the area's height