@alexpasmantier 2025-01-19
> Composers which can render a given line "statelessely" (such as the LineTruncator) which actually end up doing a lot of unnecessary work (and on the critical rendering path) when the offset gets high.
>
>I came across this while experimenting with previews in [television](https://github.com/alexpasmantier/television) and noticed increased rendering lag as the preview offset got higher for large files.

Here's an example. The slowdown is kind of subtle but I thinks it's noticeable enough to a user (even if I agree this is somewhat of an edge case).

https://github.com/user-attachments/assets/253b680a-7757-4808-a0bb-ff571d712670