@alexpasmantier 2025-01-20
@joshka Simplified the code quite a bit (finally keeping that trait) and did a bit of benchmarking with criterion using the benches [here](https://github.com/ratatui/ratatui/blob/main/ratatui/benches/main/paragraph.rs).

Found out the benches were actually broken since they were playing with the horizontal scroll when I'm pretty confident the intent was to play with the `y` scroll so I fixed that and made sure we're always rendering a comparable amount of content whatever the scroll offset is.

Here are the results (in a nutshell):

**paragraph/render_scroll_half**
|                | 64 lines  | 2048 lines | u16::MAX lines |
|----------------|:---------:|:----------:|:--------------:|
| `main`           | 292.12 µs | 3.7583 ms  | 113.54 ms      |
| `current branch` | 279.93 µs | 290.45 µs  | 973.97 µs      |
| **speedup** | NA | **-92.268%** | **-99.133%** |

**paragraph/render_scroll_full**
|                | 64 lines  | 2048 lines | u16::MAX lines |
|----------------|:---------:|:----------:|:--------------:|
| `main`           | 316.93 µs | 7.1271 ms  | 228.04 ms      |
| `current branch` | 277.21 µs | 289.69 µs  | 988.34 µs      |
| **speedup** | **-13.585%** | **-95.924%** | **-99.565%** |

All other measures remaining more or less the same.

The plot below might show in a more obvious manner how the PR makes scrolling without wrapping almost independent of the number of lines vs before.

| `main` | `this branch` |
| :--: | :--: |
| <img width="1530" alt="Screenshot 2025-01-20 at 15 07 36" src="https://github.com/user-attachments/assets/947cdbb5-2e4c-46d0-a6a7-8e16c8b635c4" /> | <img width="1507" alt="Screenshot 2025-01-20 at 15 08 34" src="https://github.com/user-attachments/assets/6c3d237b-cddb-40b6-a544-8817e847971b" /> |