@joshka 2025-01-20
Are the perf changes due to the change of the benches, or due to the change in implementation (or a combo perhaps)? Are you able to run the new benches against the code in main to double check that / show that the changes are all due to the algo change?