@alexpasmantier 2025-01-23
Any thoughts on how to proceed with these changes?
Are there any requirements you feel aren't met yet and should be improved on?