@joshka 2025-01-23
> Any thoughts on how to proceed with these changes? Are there any requirements you feel aren't met yet and should be improved on?

Added some small changes that simplify / clarify things a bit in https://github.com/ratatui/ratatui/pull/1622/commits/82f12c97016e162faca52b242cfa938c61774bd6 WDYT?